requirements.txt
slack-sdk==3.26.1
Flask==2.3.3
slackeventsapi==3.0.1
python-dotenv==1.0.0

